recombine all fields and pagination to avoid rate limits
